+++ Server/utils.h
@@ -6,7 +6,6 @@
 #include <netinet/in.h>
 #include <stdio.h>
 
-
 //color codes
 #define KRED  "\x1B[31m"
 #define KGRN  "\x1B[32m"
