+++ pub_sub_mqtt/publisher/pub.c
@@ -1,12 +1,5 @@
-// Current Assumes
-// QoS 0
-// All messages < 127 bytes
-//
-// ./mqttpub -c <client name> -i <ip address> -p <port> -t <topic> -n <loop count>
-//
-// e.g.
-// ./mqttpub -i 192.168.1.38 -t mbed/fishtank -c MacBook -n 5
-//
+// ./pub -c <client name> -i <ip address> -p <port> -t <topic> -n <loop count>
+// ./pub -i 127.0.0.1 -p 1883 -t google/gmail -c client -n 5
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -18,30 +11,30 @@
 const char *client_name = "default_pub"; 	// -c
 const char *ip_addr = "127.0.0.1";		// -i
 uint32_t port = 1883;			// -p
-const char *topic = "hello/world";	// -t //xac nhan ben nhan thong tin, co cau truc phan tang
+const char *topic = "google/gmail";	// -t //xac nhan ben nhan thong tin, co cau truc phan tang
 uint32_t count = 10;			// -n
 
 void parse_options(int argc, char** argv);
 
 int main(int argc, char** argv) {
 	puts("MQTT PUB Test Code");
-//  mqtt_broker_handle_t *broker = mqtt_connect("default_pub","127.0.0.1", 1883);
+
 	mqtt_broker_handle_t *broker = mqtt_connect(client_name, ip_addr, port);
 
 	if (broker == 0) {
 		puts("Failed to connect");
 		exit(1);
 	}
-	printf("hello\n");
+	
 	char msg[128];
 
 	for (int i = 1; i <= count; ++i) {
 		scanf("%s", msg);
 		// sprintf(msg, "Message number %d", i);
-		if (mqtt_publish(broker, topic, msg, QoS0) == -1) {
+		if (mqtt_publish(broker, topic, msg, QoS1) == -1) {
 			printf("publish failed\n");
 		} else {
-			printf("Sent %d messages\n", i);
+			//Do nothing
 		}
 		sleep(1);
 	}
